Data as Json can send and receive(without net)

+++ app/src/main/java/lucida/com/lumina/DashPoint.java
@@ -5,7 +5,7 @@ package lucida.com.lumina;
  */
 public class DashPoint
 {
-    public float left,top,right,bottom;
+    public double left,top,right,bottom;
     public int point_index;
 
     public DashPoint()
