+++ app/src/main/java/lucida/com/lumina/Point.java
@@ -4,9 +4,9 @@ package lucida.com.lumina;
  * Created by ok on 2017/4/5.
  */
 public class Point{
-    public float x,y,startx,starty;
+    public double x,y,startx,starty;
 
-    public Point(float x, float y,float startx , float starty){
+    public Point(double x, double y,double startx , double starty){
         this.x=x;
         this.y=y;
         this.startx=startx;
